+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tango-ts/migrations
 
+## 0.5.0
+
+### Patch Changes
+
+- @tango-ts/core-types@0.5.0
+- @tango-ts/orm@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes